Fix position of DEV_MAP
This rather important dictionary from constants to classes was hiding
between function definitions. The dict cannot go to the top of the file
as the classes haven't been defined there yet, so it's been pushed
to the bottom of the file.

For bonus points, the dict gets an epydoc docstring and a comment, too.